# docs/openapi.yaml
openapi: 3.1.0
info:
  title: InkFlow CRM API
  description: |
    REST API for InkFlow tattoo studio platform.

    Base path: /api (server context-path).

    Response envelope: ApiResponse<T> - see docs/API_FORMAT.md.

    Auth:
    - CRM (admin mobile / web): Supabase JWT in Authorization Bearer header.
      Custom claims: tenant_id, role, location_ids (via Supabase hook).
    - Consumer (B2C app): Supabase JWT for /public/consumer/** routes.
    - Public reads: GET /public/catalog/** and GET /public/artists/** - no token.
  contact:
    name: InkFlow
    email: dev@inkflow.studio
  license:
    name: Proprietary
  version: 1.0.0
servers:
- url: /api
  description: Current host (context-path included)
paths:
...
  /requests:
    get:
      tags:
      - CRM Â· Requests
      operationId: getAllRequests
      parameters:
      - name: pageRequest
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/PageRequest"
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: source
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: locationId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListRequestDto"
    post:
      tags:
      - CRM Â· Requests
      operationId: createRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseRequestDto"
  /requests/{id}/convert:
    post:
      tags:
      - CRM Â· Requests
      operationId: convertToClient
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConvertRequestRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseClientDto"
...
  /projects:
    get:
      tags:
      - CRM Â· Projects
      operationId: getAllProjects
      parameters:
      - name: pageRequest
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/PageRequest"
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: artistId
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
            format: uuid
      - name: clientId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: onlyMine
        in: query
        required: false
        schema:
          type: boolean
      - name: locationId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListProjectDto"
    post:
      tags:
      - CRM Â· Projects
      operationId: createProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseProjectDto"
  /projects/{id}/photos:
    post:
      tags:
      - CRM Â· Projects
      operationId: addPhoto
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddProjectPhotoRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePhotoDto"
...
  /appointments/{id}/photos:
    post:
      tags:
      - CRM Â· Appointments
      operationId: addPhoto_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddAppointmentPhotoRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePhotoDto"
...
  /requests/{id}/status:
    patch:
      tags:
      - CRM Â· Requests
      operationId: updateRequestStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequestStatusRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseRequestDto"
...
  /projects/{id}:
    get:
      tags:
      - CRM Â· Projects
      operationId: getProject
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseProjectDto"
    delete:
      tags:
      - CRM Â· Projects
      operationId: deleteProject
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
    patch:
      tags:
      - CRM Â· Projects
      operationId: updateProject
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseProjectDto"
...
  /staff/{id}/google/status:
    get:
      tags:
      - CRM Â· Integrations
      operationId: getStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseGoogleCalendarStatusResponse"
...
  /requests/{id}:
    get:
      tags:
      - CRM Â· Requests
      operationId: getRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseRequestDto"
    delete:
      tags:
      - CRM Â· Requests
      operationId: deleteRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
...
  /projects/{id}/photos/{photoId}:
    delete:
      tags:
      - CRM Â· Projects
      operationId: deletePhoto
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: photoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
...
  /appointments/{id}/photos/{photoId}:
    delete:
      tags:
      - CRM Â· Appointments
      operationId: deletePhoto_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: photoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
components:
  schemas:
...
    ErrorDto:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorDto"
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    PaginationDto:
      type: object
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
...
    CreateRequestRequest:
      type: object
      properties:
        source:
          type: string
          pattern: ^(instagram|telegram|website|referral|walk_in|other)$
        clientName:
          type: string
          maxLength: 100
          minLength: 1
        clientNickname:
          type: string
          maxLength: 50
          minLength: 0
        message:
          type: string
        phone:
          type: string
        instagram:
          type: string
      required:
      - clientName
      - source
    ApiResponseRequestDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/RequestDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    RequestDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        source:
          type: string
        clientName:
          type: string
        clientNickname:
          type: string
        message:
          type: string
        phone:
          type: string
        instagram:
          type: string
        status:
          type: string
        convertedClientId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        repliedAt:
          type: string
          format: date-time
        convertedAt:
          type: string
          format: date-time
    ConvertRequestRequest:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 50
          minLength: 1
        lastName:
          type: string
          maxLength: 50
          minLength: 1
        phone:
          type: string
          pattern: "^\\+?[0-9]{10,15}$"
        email:
          type: string
        instagram:
          type: string
        telegram:
          type: string
      required:
      - firstName
      - lastName
      - phone
    ApiResponseClientDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/ClientDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    ClientDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: string
        avatar:
          type: string
        birthDate:
          type: string
          format: date
        instagram:
          type: string
        telegram:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
        notes:
          type: string
        lastVisit:
          type: string
          format: date-time
        totalVisits:
          type: integer
          format: int32
        cancelledVisits:
          type: integer
          format: int32
        ltv:
          type: number
        createdAt:
          type: string
          format: date-time
...
    CreateProjectRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 2
        description:
          type: string
        clientId:
          type: string
          format: uuid
        artistId:
          type: string
          format: uuid
        locationId:
          type: string
          format: uuid
        estimatedCost:
          type: number
          minimum: 0.0
        totalSessions:
          type: integer
          format: int32
          minimum: 1
        sketchImage:
          type: string
      required:
      - artistId
      - clientId
      - estimatedCost
      - title
      - totalSessions
    ApiResponseProjectDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/ProjectDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    ClientSummaryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        avatar:
          type: string
        hasMedicalConditions:
          type: boolean
    PhotoDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        stage:
          type: string
        uploadedAt:
          type: string
          format: date-time
    ProjectDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        client:
          $ref: "#/components/schemas/ClientSummaryDto"
        artist:
          $ref: "#/components/schemas/StaffSummaryDto"
        status:
          type: string
        estimatedCost:
          type: number
        totalPaid:
          type: number
        totalSessions:
          type: integer
          format: int32
        completedSessions:
          type: integer
          format: int32
        sketchImage:
          type: string
        createdAt:
          type: string
          format: date-time
        photos:
          type: array
          items:
            $ref: "#/components/schemas/PhotoDto"
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/SessionDto"
    SessionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
        serviceId:
          type: string
          format: uuid
        serviceName:
          type: string
        serviceColor:
          type: string
        price:
          type: number
        finalPrice:
          type: number
        notes:
          type: string
        photosCount:
          type: integer
          format: int32
    StaffSummaryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        avatar:
          type: string
        calendarColor:
          type: string
        role:
          type: string
    AddProjectPhotoRequest:
      type: object
      properties:
        url:
          type: string
        stage:
          type: string
    ApiResponsePhotoDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/PhotoDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
...
    AddAppointmentPhotoRequest:
      type: object
      properties:
        url:
          type: string
        stage:
          type: string
      required:
      - stage
      - url
...
    UpdateRequestStatusRequest:
      type: object
      properties:
        status:
          type: string
          pattern: ^(new|replied|converted|spam)$
      required:
      - status
...
    UpdateProjectRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 2
        description:
          type: string
        artistId:
          type: string
          format: uuid
        estimatedCost:
          type: number
          minimum: 0.0
        totalSessions:
          type: integer
          format: int32
          minimum: 1
        status:
          type: string
          pattern: ^(in_progress|on_hold|completed|cancelled)$
        sketchImage:
          type: string
...
    PageRequest:
      type: object
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        sortBy:
          type: string
        sortDir:
          type: string
...
    ApiResponseGoogleCalendarStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/GoogleCalendarStatusResponse"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    GoogleCalendarStatusResponse:
      type: object
      properties:
        connected:
          type: boolean
        email:
          type: string
        configured:
          type: boolean
...
    ApiResponseListRequestDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/RequestDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
...
    ApiResponseListProjectDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProjectDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
...
  securitySchemes:
    CrmBearer:
      type: http
      description: Supabase JWT for CRM staff (OWNER/ADMIN/ARTIST).
      scheme: bearer
      bearerFormat: JWT
    ConsumerBearer:
      type: http
      description: Supabase JWT for B2C consumer app users.
      scheme: bearer
      bearerFormat: JWT